project/requirements.txt
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1.78
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.5
scikit-learn==1.3.2
pandas==2.0.3
seaborn==0.13.0
albumentations==1.3.1
efficientnet==1.1.1
requests==2.31.0
aiofiles==23.2.1
tensorflow-intel==2.13.0
typing-extensions==4.5.0